# ExplorerFrontend/app/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZondScan Public API",
    "version": "1.1.0",
...
  },
  "servers": [{ "url": "https://zondscan.com" }],
...
  "paths": {
...
    "/api/overview": {
      "get": {
        "tags": ["Overview & Market"],
        "operationId": "getOverview",
        "summary": "Network and market overview",
        "description": "Homepage hero data: market figures, wallet count, circulating supply, daily volume, validator count, and contract count in one call. Cached server side for 10 seconds.",
        "responses": {
          "200": {
            "description": "Aggregated overview.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "marketcap": { "type": "number", "description": "Market capitalization in USD." },
                    "currentPrice": { "type": "number", "description": "Spot price in USD." },
                    "priceChange24h": { "type": "number", "description": "24 hour price change in percent." },
                    "countwallets": { "type": "integer", "description": "Number of indexed wallet addresses." },
                    "circulating": { "type": "string", "description": "Circulating supply in Quanta as a decimal string." },
                    "volume": { "type": "number", "description": "Daily on-chain transferred volume." },
                    "tradingVolume": { "type": "number", "description": "24 hour trading volume in USD." },
                    "validatorCount": { "type": "integer" },
                    "contractCount": { "type": "integer" },
                    "status": {
                      "type": "object",
                      "properties": {
                        "syncing": { "type": "boolean" },
                        "dataInitialized": { "type": "boolean", "description": "False while the indexer has not yet written its first market and supply samples." }
                      }
                    }
                  }
                },
                "example": {
                  "marketcap": 32500000,
                  "currentPrice": 0.42,
                  "priceChange24h": -1.8,
                  "countwallets": 10412,
                  "circulating": "65000000",
                  "volume": 1532,
                  "tradingVolume": 210034,
                  "validatorCount": 128,
                  "contractCount": 342,
                  "status": { "syncing": true, "dataInitialized": true }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/price-history": {
      "get": {
        "tags": ["Overview & Market"],
        "operationId": "getPriceHistory",
        "summary": "Historical price samples",
        "description": "Price, market cap, and volume snapshots for charts and wallet apps.",
        "parameters": [
          {
            "name": "interval",
            "in": "query",
            "description": "Time window for the returned samples.",
            "schema": { "type": "string", "enum": ["4h", "12h", "24h", "7d", "30d", "all"], "default": "24h" }
          }
        ],
        "responses": {
          "200": {
            "description": "Price history for the requested interval.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "timestamp": { "type": "string", "description": "RFC 3339 timestamp." },
                          "priceUSD": { "type": "number" },
                          "marketCapUSD": { "type": "number" },
                          "volumeUSD": { "type": "number" }
                        }
                      }
                    },
                    "interval": { "type": "string" },
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid interval value.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/market/orderbook": {
      "get": {
        "tags": ["Overview & Market"],
        "operationId": "getMarketOrderBook",
        "summary": "Live QRL/USDT order book",
        "description": "Returns a server-side snapshot of the fixed MEXC QRLUSDT spot market: up to 100 bid levels, 100 ask levels, 100 recent trades, and 24-hour ticker data. Successful snapshots and upstream failures are cached for 3 seconds, limiting MEXC request bursts during rate limits or outages. This public endpoint requires no MEXC API key and accepts no venue, symbol, or upstream URL parameters.",
        "responses": {
          "200": {
            "description": "Normalized MEXC QRLUSDT market snapshot.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MarketOrderBook" },
                "example": {
                  "venue": "MEXC",
                  "symbol": "QRLUSDT",
                  "fetchedAt": "2026-08-09T12:14:58.123Z",
                  "lastUpdateId": 905062254,
                  "bids": [{ "price": "0.75744", "quantity": "2.17" }],
                  "asks": [{ "price": "0.76112", "quantity": "2.2" }],
                  "recentTrades": [
                    {
                      "id": "synthetic-8aa7a9881988250bd4110726-0",
                      "price": "0.76117",
                      "quantity": "11.52",
                      "quoteQuantity": "8.7686784",
                      "time": 1786290460978,
                      "aggressorSide": "buy"
                    }
                  ],
                  "ticker": {
                    "high": "0.80336",
                    "low": "0.7271",
                    "last": "0.76117",
                    "change": "-0.00088",
                    "changePercent": "-0.0011",
                    "baseVolume": "16478.3",
                    "quoteVolume": "12623.92215"
                  }
                }
              }
            }
          },
          "502": {
            "description": "MEXC returned an invalid response or could not be reached within the request budget.",
            "headers": {
              "Retry-After": {
                "description": "Seconds before ZondScan will attempt another MEXC fetch.",
                "schema": { "type": "integer", "const": 3 }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" },
                "example": { "error": "market data temporarily unavailable" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/latestblock": {
      "get": {
        "tags": ["Blocks"],
        "operationId": "getLatestBlock",
        "summary": "Latest indexed block height",
        "description": "Returns the latest synced block number as a decimal integer, plus the current QRL/USD price so fee displays need no second request. Cached server side for 5 seconds.",
        "responses": {
          "200": {
            "description": "Latest block height and spot price.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "blockNumber": { "type": "integer" },
                    "qrlUsdPrice": { "type": "number" }
                  }
                },
                "example": { "blockNumber": 148512, "qrlUsdPrice": 0.42 }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/blocks": {
      "get": {
        "tags": ["Blocks"],
        "operationId": "getBlocks",
        "summary": "Paginated block list",
        "description": "Newest blocks first, with a per-block activity rollup (token transfers and internal calls). The reported total is capped at 300 pages worth of blocks. Cached server side for 10 seconds.",
        "parameters": [
          { "name": "page", "in": "query", "description": "Page number, starting at 1.", "schema": { "type": "integer", "default": 1, "minimum": 1 } },
          { "name": "limit", "in": "query", "description": "Blocks per page. Capped at 100.", "schema": { "type": "integer", "default": 5, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "One page of blocks.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "blocks": { "type": "array", "items": { "$ref": "#/components/schemas/BlockResult" } },
                    "total": { "type": "integer", "description": "Total block count, capped at 300 pages at the current limit." },
                    "blockActivity": {
                      "type": "object",
                      "description": "Map keyed by the block's 0x-prefixed hex number. Present only for blocks with token or internal-call activity.",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "tokenTransfers": { "type": "integer" },
                          "internalCalls": { "type": "integer" }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "blocks": [
                    {
                      "number": "0x24420",
                      "hash": "0x8a4f0c2b7d1e5f3a9c6b8d0e2f4a6c8e0b2d4f6a8c0e2b4d6f8a0c2e4b6d8f0a",
                      "parentHash": "0x7b3e0d1a6c0d4e2f8b5a7c9d1e3f5a7b9d1f3a5c7e9b1d3f5a7c9e1b3d5f7a9c",
                      "timestamp": "0x6a743280",
                      "gasUsed": "0x5208",
                      "gasLimit": "0x1c9c380",
                      "baseFeePerGas": "0x3b9aca00",
                      "miner": "QaaaAAaaaAAaAaaAaAAAAaAAAaAaAaAAaAaaAaaaaAAAAAAAAaAAAAaAaAAaaAAaaaaAaAAAAaaAaAAaaaaaaAaAAaaaaAaAaaaaAaaaAAAAaAAAAaAaAaaAAAaAaaaAA",
                      "size": "0x2d0",
                      "transactions": []
                    }
                  ],
                  "total": 1500,
                  "blockActivity": { "0x24420": { "tokenTransfers": 2, "internalCalls": 1 } }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/block/{query}": {
      "get": {
        "tags": ["Blocks"],
        "operationId": "getBlock",
        "summary": "Block detail",
        "description": "Fetches one block by number, in decimal or 0x-prefixed hex form, plus per-transaction activity counts.",
        "parameters": [
          { "name": "query", "in": "path", "required": true, "description": "Block number as a decimal integer or 0x-prefixed hex string.", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Block detail with per-transaction activity map.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "block": {
                      "type": "object",
                      "properties": {
                        "jsonrpc": { "type": "string" },
                        "id": { "type": "integer" },
                        "result": { "$ref": "#/components/schemas/BlockResult" }
                      }
                    },
                    "txActivity": {
                      "type": "object",
                      "description": "Map keyed by transaction hash. Present only for transactions with token or internal-call activity.",
                      "additionalProperties": {
                        "type": "object",
                        "properties": {
                          "tokenTransfers": { "type": "integer" },
                          "internalCalls": { "type": "integer" }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "block": {
                    "jsonrpc": "2.0",
                    "id": 1,
                    "result": {
                      "number": "0x24420",
                      "hash": "0x8a4f0c2b7d1e5f3a9c6b8d0e2f4a6c8e0b2d4f6a8c0e2b4d6f8a0c2e4b6d8f0a",
                      "timestamp": "0x6a743280",
                      "gasUsed": "0x5208",
                      "gasLimit": "0x1c9c380",
                      "transactions": []
                    }
                  },
                  "txActivity": {}
                }
              }
            }
          },
          "400": {
            "description": "Invalid block number format.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "Block not found.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/api/blocksizes": {
      "get": {
        "tags": ["Blocks"],
        "operationId": "getBlockSizes",
        "summary": "Block size time series",
        "description": "Precomputed block size samples for charts, sorted by timestamp ascending and capped at 2000 points. Cached server side for 30 seconds.",
        "responses": {
          "200": {
            "description": "Block size samples.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "description": "One sample per block. blockNumber, timestamp, and size are 0x-prefixed hex strings; transactionCount is an integer.",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/txs": {
      "get": {
        "tags": ["Transactions & Mempool"],
        "operationId": "getTxs",
        "summary": "Paginated network transaction list",
        "description": "Network-wide confirmed transactions, newest first, with the total count and the latest block height for confirmation math. Cached server side for 10 seconds.",
        "parameters": [
          { "name": "page", "in": "query", "description": "Page number, starting at 1.", "schema": { "type": "integer", "default": 1, "minimum": 1 } },
          { "name": "limit", "in": "query", "description": "Items per page. Capped at 100.", "schema": { "type": "integer", "default": 10, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "One page of transactions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "txs": { "type": "array", "items": { "$ref": "#/components/schemas/TransactionByAddress" } },
                    "total": { "type": "integer" },
                    "latestBlock": { "type": "integer" }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/transactions": {
      "get": {
        "tags": ["Transactions & Mempool"],
        "operationId": "getLatestTransactions",
        "summary": "Latest transactions feed",
        "description": "Fixed-size feed of the most recent confirmed transactions, used by the homepage. Cached server side for 5 seconds.",
        "responses": {
          "200": {
            "description": "Latest transactions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": { "type": "array", "items": { "$ref": "#/components/schemas/TransactionByAddress" } }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/tx/{query}": {
      "get": {
        "tags": ["Transactions & Mempool"],
        "operationId": "getTransaction",
        "summary": "Transaction detail",
        "description": "Full detail for one confirmed transaction: transfer record, live receipt status and event logs, calldata, token transfers, internal calls, and contract metadata for decoding. This endpoint is deliberately uncached so a newly mined transaction shows its true confirmation count immediately.",
        "parameters": [
          { "name": "query", "in": "path", "required": true, "description": "Transaction hash: 0x followed by 64 hex characters.", "schema": { "type": "string", "pattern": "^0x[0-9a-fA-F]{64}$" } }
        ],
        "responses": {
          "200": {
            "description": "Transaction detail. Optional fields appear only when the transaction has the corresponding data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": { "$ref": "#/components/schemas/Transfer" },
                    "latestBlock": { "type": "integer", "description": "Latest indexed block height as a decimal integer." },
                    "receiptStatus": { "type": "string", "description": "EVM receipt status: 0x1 success, 0x0 reverted. Absent when the receipt fetch failed." },
                    "logs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "address": { "type": "string" },
                          "topics": { "type": "array", "items": { "type": "string" } },
                          "data": { "type": "string" },
                          "logIndex": { "type": "string" },
                          "removed": { "type": "boolean" },
                          "contract": { "$ref": "#/components/schemas/ContractMeta" }
                        }
                      }
                    },
                    "input": { "type": "string", "description": "Transaction calldata as a 0x-prefixed hex string. Present for contract calls." },
                    "targetContract": { "$ref": "#/components/schemas/ContractMeta" },
                    "contractCreated": {
                      "type": "object",
                      "properties": {
                        "address": { "type": "string" },
                        "isToken": { "type": "boolean" },
                        "name": { "type": "string" },
                        "symbol": { "type": "string" },
                        "decimals": { "type": "integer" },
                        "tokenStandard": { "type": "string" }
                      }
                    },
                    "tokenTransfers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "contractAddress": { "type": "string" },
                          "from": { "type": "string" },
                          "to": { "type": "string" },
                          "amount": { "type": "string" },
                          "tokenName": { "type": "string" },
                          "tokenSymbol": { "type": "string" },
                          "tokenDecimals": { "type": "integer" },
                          "tokenStandard": { "type": "string" },
                          "tokenID": { "type": "string" },
                          "logIndex": { "type": "string" }
                        }
                      }
                    },
                    "internalTransactions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "type": { "type": "string" },
                          "callType": { "type": "string" },
                          "from": { "type": "string" },
                          "to": { "type": "string" },
                          "input": { "type": "string" },
                          "output": { "type": "string" },
                          "value": { "type": "string" },
                          "gas": { "type": "string" },
                          "gasUsed": { "type": "string" },
                          "traceAddress": { "type": "array", "items": { "type": "integer" } }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "response": {
                    "ID": "665f1c2ab3d4e5f60718293a",
                    "BlockNumber": "0x24420",
                    "BlockTimestamp": "0x6a743280",
                    "From": "QaaaAAaaaAAaAaaAaAAAAaAAAaAaAaAAaAaaAaaaaAAAAAAAAaAAAAaAaAAaaAAaaaaAaAAAAaaAaAAaaaaaaAaAAaaaaAaAaaaaAaaaAAAAaAAAAaAaAaaAAAaAaaaAA",
                    "To": "QBbBbbBBbBbBBBBbBbBbBBbBBBbBBBbBbBBBbbbBBbBBbBbbBBBBbBBbbbbbBbbBBbbBBBbbBBbbbBbBBbBBbbbbBBbBBBbBbBBbbBbbBBbBBBBbbBBbbbbBBBbbBbbBb",
                    "TxHash": "0x9c1d2e3f4a5b6c7d8e9f0a1b2c3d4e5f6a7b8c9d0e1f2a3b4c5d6e7f8a9b0c1d",
                    "Value": "0x2386f26fc10000",
                    "GasUsed": "0x5208",
                    "GasPrice": "0x3b9aca00",
                    "Nonce": "0x4",
                    "Size": "0x2d0"
                  },
                  "latestBlock": 148512,
                  "receiptStatus": "0x1"
                }
              }
            }
          },
          "400": {
            "description": "Invalid transaction hash format.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "404": {
            "description": "Transaction not found.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/coinbase/{query}": {
      "get": {
        "tags": ["Transactions & Mempool"],
        "operationId": "getCoinbaseTransaction",
        "summary": "Raw transfer record lookup",
        "description": "Returns the stored transfer record for one transaction hash. A missing transaction still returns 200 with an empty record; this preserves the endpoint's historical behavior.",
        "parameters": [
          { "name": "query", "in": "path", "required": true, "description": "Transaction hash: 0x followed by 64 hex characters.", "schema": { "type": "string", "pattern": "^0x[0-9a-fA-F]{64}$" } }
        ],
        "responses": {
          "200": {
            "description": "Transfer record, empty when the hash is unknown.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "response": { "$ref": "#/components/schemas/Transfer" } }
                }
              }
            }
          },
          "400": {
            "description": "Invalid transaction hash format.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
    "/api/pending-transactions": {
      "get": {
        "tags": ["Transactions & Mempool"],
        "operationId": "getPendingTransactions",
        "summary": "Paginated mempool transactions",
        "description": "Transactions currently observed in the mempool. Cached server side for 5 seconds, matching mempool turnover.",
        "parameters": [
          { "name": "page", "in": "query", "description": "Page number, starting at 1.", "schema": { "type": "integer", "default": 1, "minimum": 1 } },
          { "name": "limit", "in": "query", "description": "Items per page. Capped at 100.", "schema": { "type": "integer", "default": 10, "maximum": 100 } }
        ],
        "responses": {
          "200": {
            "description": "One page of pending transactions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactions": { "type": "array", "items": { "$ref": "#/components/schemas/PendingTransaction" } },
                    "total": { "type": "integer" },
                    "page": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "totalPages": { "type": "integer" }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/InternalError" }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "InternalError": {
        "description": "Internal server error.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } },
        "required": ["error"]
      },
      "MarketPriceLevel": {
        "type": "object",
        "description": "One aggregated order-book price level. Decimal strings preserve exact values without floating-point conversion.",
        "properties": {
          "price": { "type": "string", "pattern": "^(?:0|[1-9][0-9]*)(?:\\.[0-9]+)?$" },
          "quantity": { "type": "string", "pattern": "^(?:0|[1-9][0-9]*)(?:\\.[0-9]+)?$" }
        },
        "required": ["price", "quantity"],
        "additionalProperties": false
      },
...
      "MarketOrderBook": {
        "type": "object",
        "properties": {
          "venue": { "type": "string", "const": "MEXC" },
          "symbol": { "type": "string", "const": "QRLUSDT" },
          "fetchedAt": { "type": "string", "format": "date-time" },
          "lastUpdateId": { "type": "integer", "format": "int64", "minimum": 0 },
          "bids": {
            "type": "array",
            "maxItems": 100,
            "description": "Bid levels in MEXC order, highest price first.",
            "items": { "$ref": "#/components/schemas/MarketPriceLevel" }
          },
          "asks": {
            "type": "array",
            "maxItems": 100,
            "description": "Ask levels in MEXC order, lowest price first.",
            "items": { "$ref": "#/components/schemas/MarketPriceLevel" }
          },
          "recentTrades": {
            "type": "array",
            "maxItems": 100,
            "items": { "$ref": "#/components/schemas/MarketTrade" }
          },
          "ticker": { "$ref": "#/components/schemas/MarketTicker" }
        },
        "required": ["venue", "symbol", "fetchedAt", "lastUpdateId", "bids", "asks", "recentTrades", "ticker"],
        "additionalProperties": false
      },
...
    }
  }
}